Reset token after resolving urls
