+++ app/src/main/java/de/koenidv/sph/networking/NetworkManager.kt
@@ -11,7 +11,6 @@ import com.androidnetworking.AndroidNetworking
 import com.androidnetworking.common.Priority
 import com.androidnetworking.error.ANError
 import com.androidnetworking.interfaces.JSONObjectRequestListener
-import com.androidnetworking.interfaces.OkHttpResponseListener
 import com.androidnetworking.interfaces.StringRequestListener
 import com.google.firebase.crashlytics.FirebaseCrashlytics
 import de.koenidv.sph.BuildConfig
@@ -23,7 +22,6 @@ import de.koenidv.sph.database.CoursesDb
 import de.koenidv.sph.database.FunctionTilesDb
 import de.koenidv.sph.debugging.DebugLog
 import de.koenidv.sph.debugging.Debugger
-import de.koenidv.sph.objects.FunctionTile
 import de.koenidv.sph.objects.FunctionTile.Companion.FEATURE_CHANGES
 import de.koenidv.sph.objects.FunctionTile.Companion.FEATURE_COURSES
 import de.koenidv.sph.objects.FunctionTile.Companion.FEATURE_MESSAGES
@@ -31,7 +29,6 @@ import de.koenidv.sph.objects.FunctionTile.Companion.FEATURE_TIMETABLE
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
-import okhttp3.Response
 import org.json.JSONObject
 import org.jsoup.Jsoup
 import java.util.*
@@ -484,98 +481,6 @@ class NetworkManager {
         }
         // Begin loading
         loadNextFeature()
-    }
-
-    /**
-     * Resolves feature urls if necessary
-     */
-    fun resolveFeatureUrls(features: List<FunctionTile>, callback: () -> Unit) {
-        // Count resolved tile urls
-        var tilesResolved = 0
-        // Only resolve sites that have this weird /index.php?a=x&e=y redirect
-        val filteredFeatures = features.filter { it.location.contains("index.php?") }
-
-        // If there are no redirects, just call back
-        if (filteredFeatures.isEmpty()) callback()
-
-        // Resolve each feature
-        for (feature in filteredFeatures) {
-            resolveUrl(feature.location, callback = { success: Int, resolved: String ->
-                // Save new url to object
-                if (success == SUCCESS || success == FAILED_UNKNOWN) {
-                    // If success or sph redirected back to home
-                    feature.location = resolved
-                }
-                // Save number of tiles resolved
-                tilesResolved++
-                // If this was the last tile
-                if (tilesResolved == filteredFeatures.size) {
-                    callback()
-                }
-            })
-        }
-    }
-
-    /**
-     * Loads an url to resolve it
-     * @return resolved url or given url if there was an error
-     */
-    private fun resolveUrl(url: String, callback: (success: Int, resolvedUrl: String) -> Unit) {
-        // Log resolving url
-        DebugLog("NetMgr", "Resolving url", bundleOf(
-                "url" to url
-        ))
-
-        // Getting an access token
-        TokenManager.authenticate { success: Int ->
-
-            if (success != SUCCESS) callback(success, url)
-
-            // Getting webpage as OkHttp
-            AndroidNetworking.get(url)
-                    .setUserAgent("koenidv/sph-planner")
-                    .setPriority(Priority.LOW)
-                    .build()
-                    .getAsOkHttpResponse(object : OkHttpResponseListener {
-                        override fun onResponse(response: Response) {
-                            // In some cases, sph redirects back to the home page (why?)
-                            // In this case, ignore the new response
-                            if (response.request().url().toString() != "https://start.schulportal.hessen.de/index.php") {
-                                callback(SUCCESS, response.request().url().toString())
-
-                                // Log success
-                                DebugLog("NetMgr", "Url resolving: success",
-                                        bundleOf(
-                                                "url" to url,
-                                                "resolvedUrl" to response.request().url().toString()
-                                        ), Debugger.LOG_TYPE_SUCCESS)
-                            } else {
-                                callback(FAILED_UNKNOWN, url)
-
-                                // Log warning
-                                DebugLog("NetMgr", "Url resolving failed",
-                                        bundleOf(
-                                                "url" to url,
-                                                "resolvedUrl" to response.request().url().toString()
-                                        ), Debugger.LOG_TYPE_WARNING)
-                            }
-                        }
-
-                        override fun onError(anError: ANError?) {
-                            // Running in emulator will cause an ssl error
-                            // Network error will also happen if there is a timeout or no connection
-                            // Not a huge deal though, the unresolved url will work as well. For now.
-                            Log.e(TAG, anError!!.errorDetail + ": " + url)
-                            callback(FAILED_UNKNOWN, url)
-
-                            // Log warning
-                            DebugLog("NetMgr", "Url resolving failed",
-                                    anError, bundleOf("url" to url),
-                                    Debugger.LOG_TYPE_WARNING)
-                        }
-
-                    })
-        }
     }
 
     /**
